+++ Civ2/Dialogs/BaseDialogHandler.cs
