+++ Civ2/Dialogs/ICivDialogHandler.cs
