+++ Civ2/Dialogs/MainMenu.cs
