+++ Civ2/Dialogs/NewGame/CustomWorldDialogs/CustomAge.cs
