+++ Civ2/Dialogs/NewGame/CustomWorldDialogs/CustomClimate.cs
