+++ Civ2/Dialogs/NewGame/CustomWorldDialogs/CustomTemp.cs
