+++ Civ2/Dialogs/NewGame/CustomWorldDialogs/CustomisePercentageLand.cs
