+++ Civ2/Dialogs/NewGame/CustomWorldSize.cs
