+++ Civ2/Dialogs/NewGame/Difficulty.cs
