+++ Civ2/Dialogs/NewGame/SelectGameVersion.cs
