+++ Civ2/Dialogs/NewGame/SimpleSettingsDialog.cs
