+++ Civ2/Dialogs/NewGame/WorldSizeHandler.cs
