+++ Civ2/Rules/Initialization.cs
