+++ Civ2/Rules/Locator.cs
