+++ Civ2clone.sln
@@ -7,13 +7,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EtoFormsUI", "EtoFormsUI\Et
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RaylibUI", "RaylibUI\RaylibUI.csproj", "{091F1A96-298E-4657-B14B-74D477EC4CF6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Civ", "Civ\Civ.csproj", "{C1C3D7F0-7D12-4C5B-A8E0-F80C3EB1F655}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Civ2", "Civ2\Civ2.csproj", "{C1C3D7F0-7D12-4C5B-A8E0-F80C3EB1F655}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Model", "Model\Model.csproj", "{F8B6C708-6732-4888-BEF0-40F83D432C01}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Civ2", "Civ2\Civ2.csproj", "{E9BE038E-8261-42D5-8801-BDCD1139080A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Civ2Gold", "Civ2Gold\Civ2Gold.csproj", "{E9BE038E-8261-42D5-8801-BDCD1139080A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TOT", "TOT\TOT.csproj", "{7DB48D41-7603-4F4A-9F06-B961649AB7E4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Civ2TOT", "Civ2TOT\Civ2TOT.csproj", "{7DB48D41-7603-4F4A-9F06-B961649AB7E4}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "Engine\Core.csproj", "{51609735-D47F-4A91-B4B4-EEED8879A3A1}"
 EndProject
