--- CivAxx.sln
@@ -1,42 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CivAxx.Gtk", "CivAxx\CivAxx.Gtk\CivAxx.Gtk.csproj", "{E52E1ACC-12A2-4989-ABBE-45C841088EA5}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CivAxx.Mac", "EtoFormsUI\CivAxx.Mac\CivAxx.Mac.csproj", "{0D2E1EC1-1BC3-4111-8D5C-CAE3C443C12C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "Engine\Core.csproj", "{891DEEF2-02E8-4E12-A654-00AB9C01A8B0}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EtoFormsUI", "EtoFormsUI\EtoFormsUI\EtoFormsUI.csproj", "{23F1E3DB-9C9C-4CA1-ACF9-028308C69995}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Civ", "Civ\Civ.csproj", "{815CC02D-AD3E-4BA2-9A26-38212BE3F5CE}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E52E1ACC-12A2-4989-ABBE-45C841088EA5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E52E1ACC-12A2-4989-ABBE-45C841088EA5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E52E1ACC-12A2-4989-ABBE-45C841088EA5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E52E1ACC-12A2-4989-ABBE-45C841088EA5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0D2E1EC1-1BC3-4111-8D5C-CAE3C443C12C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0D2E1EC1-1BC3-4111-8D5C-CAE3C443C12C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0D2E1EC1-1BC3-4111-8D5C-CAE3C443C12C}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
-		{0D2E1EC1-1BC3-4111-8D5C-CAE3C443C12C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0D2E1EC1-1BC3-4111-8D5C-CAE3C443C12C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0D2E1EC1-1BC3-4111-8D5C-CAE3C443C12C}.Release|Any CPU.Deploy.0 = Release|Any CPU
-		{891DEEF2-02E8-4E12-A654-00AB9C01A8B0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{891DEEF2-02E8-4E12-A654-00AB9C01A8B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{891DEEF2-02E8-4E12-A654-00AB9C01A8B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{891DEEF2-02E8-4E12-A654-00AB9C01A8B0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{23F1E3DB-9C9C-4CA1-ACF9-028308C69995}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{23F1E3DB-9C9C-4CA1-ACF9-028308C69995}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{23F1E3DB-9C9C-4CA1-ACF9-028308C69995}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{23F1E3DB-9C9C-4CA1-ACF9-028308C69995}.Release|Any CPU.Build.0 = Release|Any CPU
-		{815CC02D-AD3E-4BA2-9A26-38212BE3F5CE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{815CC02D-AD3E-4BA2-9A26-38212BE3F5CE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{815CC02D-AD3E-4BA2-9A26-38212BE3F5CE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{815CC02D-AD3E-4BA2-9A26-38212BE3F5CE}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
