+++ RaylibUI/Bitmaps/Images.ImportBitmaps.cs
@@ -42,7 +42,7 @@ namespace RaylibUI
         //    }
         //    catch
         //    {
-        //        Debug.WriteLine("Civ2.exe not found!");
+        //        Debug.WriteLine("Civ2Gold.exe not found!");
         //    }
         //}
